+++ tests/archetype-assembly/run.sh
@@ -2,17 +2,32 @@
 
 set -e
 
+# Clean previous project
+rm -rf quickstart-test
+
+# Get current Flowman version
+FLOWMAN_VERSION=$(mvn -f ../.. -q -N help:evaluate -Dexpression=project.version -DforceStdout)
+
 mvn archetype:generate \
     -B \
     -DarchetypeGroupId=com.dimajix.flowman.maven \
     -DarchetypeArtifactId=flowman-archetype-assembly \
     -DgroupId=test \
-    -DartifactId=quickstart \
-    -Dversion=1.0-SNAPSHOT \
-    -DflowmanVersion=1.0.0-SNAPSHOT
+    -DartifactId=quickstart-test \
+    -Dversion=1.0-SNAPSHOT
 
-cd quickstart || exit
+# Replace Flowman version
+xmlstarlet ed \
+  --inplace \
+  -N x=http://maven.apache.org/POM/4.0.0 \
+  --update /x:project/x:parent/x:version \
+  --value "$FLOWMAN_VERSION" \
+  quickstart-test/pom.xml
+
+cd quickstart-test || exit
 mvn clean install
 
 cd ..
-rm -rf quickstart
+
+# Clean up
+rm -rf quickstart-test