+++ docs/releases.md
@@ -101,6 +101,7 @@ changes over time.
 * github-404: [BUG] Partition columns do not support Timestamp data type
 * github-409: [BUG] Fix build for AWS EMR 6.10 and Azure Synapse 3.3
 * github-407: Update Delta to 2.3.0 for Spark 3.3
+* github-406: Improve integration tests to automatically pick up the current Flowman version
 
 #### Breaking changes